+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## master (unreleased)
-
+* Adds Ahoy generator. ([@abhaynikam][])
 ## 0.8.0 (February 28th, 2021)
 * Adds Devise Google Omniauth generator. ([@abhaynikam][])
 * Adds Devise Twitter Omniauth generator. ([@abhaynikam][])